+++ functions.php
@@ -40,3 +40,5 @@ require get_template_directory() . '/inc/performance.php';
 require get_template_directory() . '/inc/security.php';
 
 require get_template_directory() . '/inc/schema.php';
+
+require get_template_directory() . '/inc/social-image.php';